Refactor assertions to helper file
